+++ R/interpret_status.R
@@ -1,25 +1,29 @@
-#' @title Interpret a set of package issues
+#' @title Interpret the status of a package
 #' @export
-#' @family issues
+#' @family status
 #' @description Summarize the issues of a package in human-readable text.
 #' @return A character string summarizing the issues of a package in prose.
-#' @param path Character string, file path to a JSON issue file
-#'   of a package.
-interpret_issue <- function(path) {
-  package <- basename(path)
-  if (!file.exists(path)) {
+#' @param package Character string, name of the package.
+#' @param issues A list with one issue per package. Obtained by
+#'   reading the results of [record_issues()].
+interpret_status <- function(package, issues) {
+  if (is.null(issues[[package]])) {
     return(paste("Package", package, "has no recorded issues."))
   }
-  issue <- jsonlite::read_json(path, simplifyVector = TRUE)
-  paste0(
-    interpret_title(issue, package),
-    interpret_advisories(issue),
-    interpret_checks(issue),
-    interpret_dependencies(issue, package),
-    interpret_licenses(issue, package),
-    interpret_remotes(issue),
-    interpret_versions(issue)
+  issue <- issues[[package]]
+  out <- paste(
+    c(
+      interpret_title(issue, package),
+      interpret_advisories(issue),
+      interpret_checks(issue),
+      interpret_dependencies(issue, package),
+      interpret_licenses(issue, package),
+      interpret_remotes(issue),
+      interpret_versions(issue)
+    ),
+    collapse = ""
   )
+  trimws(out)
 }
 
 interpret_title <- function(issue, package) {
@@ -33,7 +37,7 @@ interpret_title <- function(issue, package) {
   if (is.character(issue$remote_hash)) {
     title <- paste(title, "remote hash", issue$remote_hash)
   }
-  paste0(title, " on ", issue$date, ".")
+  paste0(title, " since ", issue$date, ".")
 }
 
 interpret_advisories <- function(issue) {